website/requirements.txt
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.2
asn1crypto==0.24.0
astroid==2.2.5
astropy==3.1.2
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.7.1
bitarray==0.8.3
bkcharts==0.2
bleach==3.1.0
bokeh==1.0.4
boto==2.49.0
Bottleneck==1.2.1
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cloudpickle==0.8.0
clyent==1.2.2
colorama==0.4.1
colors.py==0.2.2
colour==0.1.5
conda==4.6.11
conda-build==3.17.8
conda-verify==3.1.1
contentful==1.12.3
contextlib2==0.5.5
cryptography==2.6.1
cycler==0.10.0
Cython==0.29.6
cytoolz==0.9.0.1
dask==1.1.4
decorator==4.4.0
defusedxml==0.6.0
diff-match-patch==20181111
distributed==1.26.0
Django==2.2.1
django-ckeditor==5.8.0
django-countries==5.5
django-extra-views==0.11.0
django-facebook-comments==0.1.5
django-haystack==2.8.1
django-import-export==1.2.0
django-js-asset==1.2.2
django-multiselectfield==0.1.9
django-oscar==2.0.3
django-oscar-catalogue==0.0.1
django-phonenumber-field==2.0.1
django-social-share==1.3.2
django-summernote==0.8.11.4
django-tables2==1.21.2
django-treebeard==4.3
django-widget-tweaks==1.4.5
docutils==0.14
entrypoints==0.3
et-xmlfile==1.0.1
factory-boy==2.12.0
Faker==2.0.3
fastcache==1.0.2
filelock==3.0.10
Flask==1.0.2
future==0.17.1
gevent==1.4.0
glob2==0.6
gmpy2==2.0.8
greenlet==0.4.15
gTTS==2.0.4
gTTS-token==1.1.3
h5py==2.9.0
heapdict==1.0.0
html5lib==1.0.1
idna==2.8
imageio==2.6.1
imagesize==1.1.0
IMAPClient==2.1.0
importlib-metadata==0.0.0
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.20
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.13.3
jeepney==0.4
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyter-tabnine==1.0.2
jupyterlab==0.35.4
jupyterlab-server==0.2.0
keyring==18.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
libarchive-c==2.8
lief==0.9.0
llvmlite==0.28.0
locket==0.2.0
lxml==4.3.3
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.10
mkl-random==1.0.2
more-itertools==6.0.0
MouseInfo==0.0.4
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.3
nltk==3.4
nose==1.3.7
notebook==5.7.8
numba==0.43.1
numexpr==2.6.9
numpy==1.16.3
numpydoc==0.8.0
odfpy==1.4.0
olefile==0.46
opencv-contrib-python==4.1.1.26
opencv-python==4.1.1.26
openpyxl==3.0.0
packaging==19.0
panda==0.3.1
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.4
partd==0.3.10
path.py==11.5.0
pathlib2==2.3.3
patsy==0.5.1
pep8==1.7.1
pexpect==4.6.0
phonenumbers==8.10.22
pickleshare==0.7.5
Pillow==5.4.1
pkginfo==1.5.0.1
pluggy==0.9.0
ply==3.11
prometheus-client==0.6.0
prompt-toolkit==2.0.9
psutil==5.6.1
ptyprocess==0.6.0
purl==1.5
py==1.8.0
PyAutoGUI==0.9.47
pycodestyle==2.5.0
pycosat==0.6.3
pycountry==19.8.18
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.2
pyflakes==2.1.1
pygame==1.9.6
PyGetWindow==0.0.7
Pygments==2.3.1
pyimage==0.1.1
pylint==2.3.1
PyMsgBox==1.0.7
pyodbc==4.0.26
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyperclip==1.7.0
PyRect==0.1.4
pyrsistent==0.14.11
PyScreeze==0.1.22
PySocks==1.6.8
pytest==4.3.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.3.0
pytest-openfiles==0.3.2
pytest-remotedata==0.3.1
python-dateutil==2.8.0
python3-openid==3.1.0
python3-xlib==0.15
PyTweening==1.0.3
pytz==2018.9
PyWavelets==1.0.3
PyYAML==5.1
pyzmq==18.0.0
QtAwesome==0.5.7
qtconsole==4.4.3
QtPy==1.7.0
regex==2019.4.14
requests==2.21.0
requests-oauthlib==1.3.0
rope==0.12.0
ruamel-yaml==0.15.46
ruamel.yaml==0.15.96
scikit-image==0.15.0
scikit-learn==0.20.3
scipy==1.3.0
seaborn==0.9.0
SecretStorage==3.1.1
secure-smtplib==0.1.1
selenium==3.141.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
snowballstemmer==1.2.1
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sorl-thumbnail==12.4.1
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.8
SpeechRecognition==3.8.1
Sphinx==1.8.5
sphinxcontrib-websupport==1.1.0
spyder==3.3.3
spyder-kernels==0.4.2
SQLAlchemy==1.3.1
sqlparse==0.3.0
statsmodels==0.9.0
sympy==1.3
tables==3.5.1
tablib==0.14.0
tblib==1.3.2
terminado==0.8.1
testpath==0.4.2
text-unidecode==1.3
toolz==0.9.0
tornado==6.0.2
tqdm==4.31.1
traitlets==4.3.2
twilio==6.33.0
typed-ast==1.3.5
unicodecsv==0.14.1
urllib3==1.24.1
virtualenv==16.7.7
wcwidth==0.1.7
webcolors==1.10
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.11.1
wurlitzer==1.0.2
xlrd==1.2.0
XlsxWriter==1.1.5
xlwt==1.3.0
zict==0.1.4
zipp==0.3.3
